pyproject.toml
[project]
name="jirascripting"
version='0.0.4'
description='Methods for interacting with Jira Cloud'
authors = [
  { name="David Renton", email="davidleerenton@gmail.com" },
]
dependencies=["requests~=2.31.0", "python-dotenv~=0.19.0"]
classifiers=[
    "Development Status :: 3 - Alpha",
    'License :: OSI Approved :: MIT License',
    "Programming Language :: Python :: 3",
    'Operating System :: OS Independent',
]

